+++ migrations/028_store_last_account_error.rb
@@ -1,0 +1,11 @@
+Sequel.migration do
+  up do
+    add_column :accounts, :last_error, String
+    add_column :accounts, :last_error_at, DateTime
+  end
+
+  down do
+    drop_column :accounts, :last_error_at
+    drop_column :accounts, :last_error
+  end
+end