+++ spec/epics/box/models/account_spec.rb
@@ -16,8 +16,12 @@ module Epics
         context 'no activated subscriber available' do
           before { account.add_subscriber(activated_at: nil) }
 
-          it 'fails with an no transport client exceptiopn' do
+          it 'fails with an no transport client exception' do
             expect { account.transport_client }.to raise_error(Account::NoTransportClient)
+          end
+          it 'persists error in database' do
+            expect { account.transport_client rescue nil }.to change { account.reload.last_error }.
+              from(nil).to('Please setup and activate at least one subscriber with a transport signature')
           end
         end